+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - This CHANGELOG file to hopefully capture any change notes
 - `evaluate` method to public interface.
 - Code Climate badge to README.
+- `$group` stage and accumulators.
+- Custom accumulator support via `accumulator` method.
+- `defaultAccumulators` initialization option.
 
 ### Changed
 - The public interface into a type accessible via `<instance>.Critr`